# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: SkillPath API
  version: v1
  description: SkillPath identity, goal, graph, assessment, progress, and review contracts through Phase 4.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/goal-templates:
    get:
      operationId: listGoalTemplates
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Active goal templates.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalTemplateResponse'
  /api/v1/goal-templates/{goalTemplateId}/graph:
    get:
      operationId: getPublishedGoalGraph
      parameters:
        - $ref: '#/components/parameters/GoalTemplateId'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: anchorNodeId
          in: query
          schema: { type: string }
        - name: depth
          in: query
          schema: { type: integer, minimum: 0, maximum: 10, default: 2 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Current published graph for the active goal template.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalGraphResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
  /api/v1/goals:
    post:
      operationId: createGoal
      security:
        - cookieSession: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '200':
          description: Original result replayed for the same idempotency key and request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/goals/active:
    get:
      operationId: getActiveGoal
      security:
        - cookieSession: []
      responses:
        '200':
          description: Current active goal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/assessments/diagnostic:
    post:
      operationId: startDiagnostic
      security:
        - cookieSession: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '201':
          description: New diagnostic session created with pinned question versions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSessionResponse'
        '200':
          description: Existing in-progress or completed diagnostic returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSessionResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/assessments/{sessionId}/next-question:
    get:
      operationId: getNextDiagnosticQuestion
      security:
        - cookieSession: []
      parameters:
        - $ref: '#/components/parameters/AssessmentSessionId'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Next pinned unanswered question. The answer key is never included.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentQuestionResponse'
        '204':
          description: Diagnostic is completed and has no next question.
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
  /api/v1/assessments/{sessionId}/attempts:
    post:
      operationId: submitDiagnosticAttempt
      security:
        - cookieSession: []
      parameters:
        - $ref: '#/components/parameters/AssessmentSessionId'
        - $ref: '#/components/parameters/CsrfHeader'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAssessmentAttemptRequest'
      responses:
        '201':
          description: Attempt, concept evidence, and durable handoff created atomically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentAttemptResponse'
        '200':
          description: Original accepted attempt replayed for the same key and payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentAttemptResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/assessments/{sessionId}/result:
    get:
      operationId: getDiagnosticResult
      security:
        - cookieSession: []
      parameters:
        - $ref: '#/components/parameters/AssessmentSessionId'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Observational concept evidence, not authoritative mastery or a recommendation.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResultResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: SKILLPATH_SESSION
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: Preferred learner presentation language. Unsupported values fall back to English.
      schema:
        type: string
        enum: [vi-VN, en]
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
    GoalTemplateId:
      name: goalTemplateId
      in: path
      required: true
      schema: { type: string }
...
    AssessmentSessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string }
...
  headers:
    ContentLanguage:
      description: Resolved language used for localized learner content.
      schema:
        type: string
        enum: [vi-VN, en]
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
  schemas:
...
    GoalTemplateResponse:
      type: object
      required: [id, key, displayName, description]
      properties:
        id: { type: string }
        key: { type: string }
        displayName: { type: string }
        description: { type: string }
    CreateGoalRequest:
      type: object
      required: [goalTemplateId, targetDate, timezone, defaultDailyMinutes]
      properties:
        goalTemplateId: { type: string }
        targetDate: { type: string, format: date }
        timezone: { type: string, maxLength: 64 }
        defaultDailyMinutes: { type: integer, minimum: 30, maximum: 180 }
    GoalResponse:
      type: object
      required: [id, goalTemplateId, targetDate, timezone, defaultDailyMinutes, status, version]
      properties:
        id: { type: string }
        goalTemplateId: { type: string }
        targetDate: { type: string, format: date }
        timezone: { type: string }
        defaultDailyMinutes: { type: integer }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
        version: { type: integer, format: int64 }
...
    GoalGraphNode:
      type: object
      required: [id, slug, name, description, category, difficulty, estimatedMinutes, status, relevanceWeight, requiredMastery, terminal]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        difficulty: { type: integer }
        estimatedMinutes: { type: integer }
        status: { type: string }
        relevanceWeight: { type: number, minimum: 0, maximum: 1 }
        requiredMastery: { type: number, minimum: 0, maximum: 1 }
        terminal: { type: boolean }
    KnowledgeEdge:
      type: object
      required: [id, sourceNodeId, targetNodeId, type, strength, status, rationale]
      properties:
        id: { type: string }
        sourceNodeId: { type: string }
        targetNodeId: { type: string }
        type: { type: string, enum: [PREREQUISITE, PART_OF, RELATED, APPLIED_IN] }
        strength: { type: number, minimum: 0, maximum: 1 }
        status: { type: string, enum: [ACTIVE, DEPRECATED] }
        rationale: { type: string }
    GoalGraphResponse:
      type: object
      required: [graphVersionId, curriculumKey, versionLabel, publishedAt, nodes, edges, truncated]
      properties:
        graphVersionId: { type: string }
        curriculumKey: { type: string }
        versionLabel: { type: string }
        publishedAt: { type: string, format: date-time }
        nodes:
          type: array
          items: { $ref: '#/components/schemas/GoalGraphNode' }
        edges:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeEdge' }
        truncated: { type: boolean }
        nextCursor: { type: [string, 'null'] }
...
    AssessmentSessionResponse:
      type: object
      required: [id, goalId, graphVersionId, assessmentPolicyVersion, status, startedAt, expiresAt, answeredQuestions, totalQuestions, created, resumed]
      properties:
        id: { type: string }
        goalId: { type: string }
        graphVersionId: { type: string }
        assessmentPolicyVersion: { type: string }
        status: { type: string, enum: [IN_PROGRESS, COMPLETED, EXPIRED] }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        answeredQuestions: { type: integer, minimum: 0 }
        totalQuestions: { type: integer, minimum: 0 }
        created: { type: boolean }
        resumed: { type: boolean }
    AssessmentQuestionOption:
      type: object
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
    AssessmentQuestionResponse:
      type: object
      required: [sessionId, sessionQuestionId, questionVersionId, type, prompt, difficulty, estimatedSeconds, position, totalQuestions, expiresAt, options]
      properties:
        sessionId: { type: string }
        sessionQuestionId: { type: string }
        questionVersionId: { type: string }
        type: { type: string, enum: [SINGLE_CHOICE, MULTIPLE_CHOICE] }
        prompt: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        estimatedSeconds: { type: integer, minimum: 1 }
        position: { type: integer, minimum: 1 }
        totalQuestions: { type: integer, minimum: 1 }
        expiresAt: { type: string, format: date-time }
        options:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/AssessmentQuestionOption' }
    SubmitAssessmentAttemptRequest:
      type: object
      required: [sessionQuestionId, selectedOptionIds, timeSpentSeconds]
      properties:
        sessionQuestionId: { type: string }
        selectedOptionIds:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items: { type: string, maxLength: 120 }
        selfConfidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        timeSpentSeconds: { type: integer, minimum: 0, maximum: 3600 }
    AssessmentAttemptResponse:
      type: object
      required: [attemptId, sessionStatus, replayed]
      properties:
        attemptId: { type: string }
        sessionStatus: { type: string, enum: [IN_PROGRESS, COMPLETED, EXPIRED] }
        replayed: { type: boolean }
    AssessmentEvidenceResponse:
      type: object
      required: [evidenceId, attemptId, knowledgeNodeId, knowledgeNodeSlug, knowledgeNodeName, dimension, score, reliability, evaluatorVersion, observedAt]
      description: Observational evidence from one evaluated attempt; not authoritative mastery.
      properties:
        evidenceId: { type: string }
        attemptId: { type: string }
        knowledgeNodeId: { type: string }
        knowledgeNodeSlug: { type: string }
        knowledgeNodeName: { type: string }
        dimension: { type: string, enum: [RECOGNITION, UNDERSTANDING] }
        score: { type: number, minimum: 0, maximum: 1 }
        reliability: { type: number, minimum: 0, maximum: 1 }
        evaluatorVersion: { type: string }
        observedAt: { type: string, format: date-time }
    AssessmentResultResponse:
      type: object
      required: [sessionId, graphVersionId, assessmentPolicyVersion, startedAt, completedAt, answeredQuestions, totalQuestions, overallObjectiveScore, evidence, interpretation]
      properties:
        sessionId: { type: string }
        graphVersionId: { type: string }
        assessmentPolicyVersion: { type: string }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        answeredQuestions: { type: integer, minimum: 0 }
        totalQuestions: { type: integer, minimum: 0 }
        overallObjectiveScore: { type: number, minimum: 0, maximum: 1 }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/AssessmentEvidenceResponse' }
        interpretation: { type: string }
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
    ApiProblem:
      type: object
      required: [type, title, status, code, detail, instance, correlationId, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...